restructure the packages so that base classes could be reused
