+++ tests/tests-framework/src/main/java/org/eclipse/dirigible/tests/projects/BaseTestProject.java
@@ -7,7 +7,7 @@
  *
  * SPDX-FileCopyrightText: Eclipse Dirigible contributors SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.dirigible.integration.tests.ui.tests.projects;
+package org.eclipse.dirigible.tests.projects;
 
 import org.eclipse.dirigible.tests.EdmView;
 import org.eclipse.dirigible.tests.FormView;
