+++ tests/tests-integrations/src/test/java/org/eclipse/dirigible/integration/tests/api/rest/TenantDeterminationIT.java
@@ -11,8 +11,8 @@ package org.eclipse.dirigible.integration.tests.api.rest;
 
 import io.restassured.RestAssured;
 import io.restassured.specification.RequestSpecification;
-import org.eclipse.dirigible.integration.tests.IntegrationTest;
-import org.eclipse.dirigible.integration.tests.TenantCreator;
+import org.eclipse.dirigible.tests.IntegrationTest;
+import org.eclipse.dirigible.tests.TenantCreator;
 import org.eclipse.dirigible.tests.restassured.RestAssuredExecutor;
 import org.hamcrest.Matchers;
 import org.springframework.beans.factory.annotation.Autowired;
