+++ tests/tests-integrations/src/test/java/org/eclipse/dirigible/integration/tests/ui/tests/ApproveLeaveRequestBpmIT.java
@@ -9,9 +9,9 @@
  */
 package org.eclipse.dirigible.integration.tests.ui.tests;
 
-import org.eclipse.dirigible.integration.tests.ui.tests.projects.PredefinedProjectIT;
-import org.eclipse.dirigible.integration.tests.ui.tests.projects.TestProject;
 import org.eclipse.dirigible.tests.mail.GreenMailConfig;
+import org.eclipse.dirigible.tests.projects.PredefinedProjectIT;
+import org.eclipse.dirigible.tests.projects.TestProject;
 import org.springframework.beans.factory.annotation.Autowired;
 
 class ApproveLeaveRequestBpmIT extends PredefinedProjectIT {
