+++ tests/tests-integrations/src/test/java/org/eclipse/dirigible/integration/tests/ui/tests/BPMLeaveRequestTestProject.java
@@ -12,7 +12,6 @@ package org.eclipse.dirigible.integration.tests.ui.tests;
 import com.icegreen.greenmail.util.GreenMail;
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.MimeMessage;
-import org.eclipse.dirigible.integration.tests.ui.tests.projects.BaseTestProject;
 import org.eclipse.dirigible.tests.EdmView;
 import org.eclipse.dirigible.tests.IDE;
 import org.eclipse.dirigible.tests.IDEFactory;
@@ -21,6 +20,7 @@ import org.eclipse.dirigible.tests.framework.HtmlElementType;
 import org.eclipse.dirigible.tests.mail.EmailAsserter;
 import org.eclipse.dirigible.tests.mail.EmailAssertion;
 import org.eclipse.dirigible.tests.mail.EmailAssertionBuilder;
+import org.eclipse.dirigible.tests.projects.BaseTestProject;
 import org.eclipse.dirigible.tests.util.ProjectUtil;
 import org.eclipse.dirigible.tests.util.SecurityUtil;
 import org.springframework.context.annotation.Lazy;
