+++ tests/tests-integrations/src/test/java/org/eclipse/dirigible/integration/tests/ui/tests/DependsOnIT.java
@@ -9,8 +9,8 @@
  */
 package org.eclipse.dirigible.integration.tests.ui.tests;
 
-import org.eclipse.dirigible.integration.tests.ui.tests.projects.PredefinedProjectIT;
-import org.eclipse.dirigible.integration.tests.ui.tests.projects.TestProject;
+import org.eclipse.dirigible.tests.projects.PredefinedProjectIT;
+import org.eclipse.dirigible.tests.projects.TestProject;
 import org.springframework.beans.factory.annotation.Autowired;
 
 class DependsOnIT extends PredefinedProjectIT {
