+++ tests/tests-integrations/src/test/java/org/eclipse/dirigible/integration/tests/ui/tests/DirigibleHomepageIT.java
@@ -9,6 +9,7 @@
  */
 package org.eclipse.dirigible.integration.tests.ui.tests;
 
+import org.eclipse.dirigible.tests.UserInterfaceIntegrationTest;
 import org.eclipse.dirigible.tests.framework.HtmlElementType;
 import org.junit.jupiter.api.Test;
