+++ tests/tests-integrations/src/test/java/org/eclipse/dirigible/integration/tests/ui/tests/OrderedTestSuite.java
@@ -9,15 +9,13 @@
  */
 package org.eclipse.dirigible.integration.tests.ui.tests;
 
-import org.eclipse.dirigible.integration.tests.api.SecurityIT;
-import org.eclipse.dirigible.integration.tests.api.java.messaging.MessagingFacadeIT;
 import org.junit.platform.suite.api.SelectClasses;
 import org.junit.platform.suite.api.Suite;
 import org.junit.platform.suite.api.SuiteDisplayName;
 
 @Suite
 @SuiteDisplayName("Ordered Test Suite")
-@SelectClasses({SecurityIT.class, MessagingFacadeIT.class, CsvimIT.class})
+@SelectClasses({ApproveLeaveRequestBpmIT.class, DeclineLeaveRequestBpmIT.class})
 public class OrderedTestSuite {
     // use this suite class to run tests in specific order if needed
     // it is not configured to be executed automatically by the maven plugins
